+++ lib/config/default.config.ts
@@ -426,6 +426,19 @@ const defaultConfig: KuzzleConfiguration = {
               },
             },
           },
+          imports: {
+            settings: {
+              // @deprecated : replace undefined by 1
+              number_of_shards: 1,
+              number_of_replicas: 1,
+            },
+            mappings: {
+              dynamic: "strict",
+              properties: {
+                hash: { type: "keyword" },
+              },
+            },
+          },
         },
       },
       maxScrollDuration: "1m",