+++ lib/core/plugin/pluginContext.ts
@@ -23,22 +23,21 @@ import Bluebird from "bluebird";
 import { Koncorde } from "../shared/KoncordeWrapper";
 import { JSONObject } from "kuzzle-sdk";
 
-import { EmbeddedSDK } from "../shared/sdk/embeddedSdk";
-import PluginRepository from "./pluginRepository";
-import Store from "../shared/store";
-import { Elasticsearch } from "../../service/storage/Elasticsearch";
-import { isPlainObject } from "../../util/safeObject";
-import Promback from "../../util/promback";
-import { Mutex } from "../../util/mutex";
+
+import {
+  KuzzleRequest,
+  Request,
+  RequestContext,
+  RequestInput,
+} from "../../../index";
 import * as kerror from "../../kerror";
-import storeScopeEnum from "../storage/storeScopeEnum";
 import {
   BadRequestError,
   ExternalServiceError,
   ForbiddenError,
   GatewayTimeoutError,
-  InternalError as KuzzleInternalError,
   KuzzleError,
+  InternalError as KuzzleInternalError,
   NotFoundError,
   PartialError,
   PluginImplementationError,
@@ -48,13 +47,15 @@ import {
   TooManyRequestsError,
   UnauthorizedError,
 } from "../../kerror/errors";
-import {
-  RequestContext,
-  RequestInput,
-  KuzzleRequest,
-  Request,
-} from "../../../index";
+import { Elasticsearch } from "../../service/storage/elasticsearch";
+import { Mutex } from "../../util/mutex";
+import Promback from "../../util/promback";
+import { isPlainObject } from "../../util/safeObject";
 import { BackendCluster } from "../backend";
+import { EmbeddedSDK } from "../shared/sdk/embeddedSdk";
+import { Store } from "../shared/store";
+import { storeScopeEnum } from "../storage/storeScopeEnum";
+import PluginRepository from "./pluginRepository";
 
 const contextError = kerror.wrap("plugin", "context");