+++ lib/service/storage/7/elasticsearch.ts
@@ -26,16 +26,16 @@ import { Index, IndicesCreate } from "sdk-es7/api/requestParams";
 import { TypeMapping } from "sdk-es7/api/types";
 import {
   InfoResult,
-  KRequestBody,
   JSONObject,
   KImportError,
+  KRequestBody,
   KRequestParams,
 } from "../../../types/storage/7/Elasticsearch";
 
 import assert from "assert";
 
-import ms from "ms";
 import Bluebird from "bluebird";
+import ms from "ms";
 import semver from "semver";
 import debug from "../../../util/debug";
 
@@ -45,10 +45,10 @@ import didYouMean from "../../../util/didYouMean";
 import * as kerror from "../../../kerror";
 import { assertIsObject } from "../../../util/requestAssertions";
 import { isPlainObject } from "../../../util/safeObject";
-import scopeEnum from "../../../core/storage/storeScopeEnum";
 import extractFields from "../../../util/extractFields";
 import { Mutex } from "../../../util/mutex";
 import { randomNumber } from "../../../util/name-generator";
+import { storeScopeEnum } from "../../../core/storage/storeScopeEnum";
 
 debug("kuzzle:services:elasticsearch");
 
@@ -90,7 +90,7 @@ let esState = esStateEnum.NONE;
  */
 export class ES7 {
   public _client: Client;
-  public _scope: scopeEnum;
+  public _scope: storeScopeEnum;
   public _indexPrefix: string;
   public _esWrapper: ESWrapper;
   public _esVersion: any;
@@ -102,11 +102,11 @@ export class ES7 {
   public scrollTTL: number;
   public _config: any;
 
-  constructor(config: any, scope = scopeEnum.PUBLIC) {
+  constructor(config: any, scope = storeScopeEnum.PUBLIC) {
     this._config = config;
     this._scope = scope;
     this._indexPrefix =
-      scope === scopeEnum.PRIVATE ? PRIVATE_PREFIX : PUBLIC_PREFIX;
+      scope === storeScopeEnum.PRIVATE ? PRIVATE_PREFIX : PUBLIC_PREFIX;
 
     this._client = null;
     this._esWrapper = null;