+++ lib/types/shared/StoreCollectionsDefinition.ts
@@ -1,0 +1,17 @@
+import { JSONObject } from "../../../index";
+
+export interface StoreCollectionDefinition {
+  /**
+   * Collection mappings
+   */
+  mappings: JSONObject;
+  /**
+   * Collection settings
+   */
+  settings?: JSONObject;
+}
+
+export type StoreCollectionsDefinition = Record<
+  string,
+  StoreCollectionDefinition
+>;